complete model part of orderSummary
